chore: refactor curves
