+++ packages/core/src/contrib/Curves.ts
